# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: TMS Tyre Service API
  description: API para la gestión de neumáticos y observaciones en el sistema TMS
  version: 1.0.0
  contact:
    name: DLIRIO Tech Team
    email: support@dliriotech.com
servers:
  - url: http://localhost:8080
    description: Desarrollo local
  - url: https://api.tms.dliriotech.com
    description: Producción
...
paths:
...
  /api/v1/observaciones-neumaticos/neumatico/{neumaticoId}/solucionables:
    get:
      tags:
        - observaciones
      summary: Obtener observaciones solucionables
      description: Retorna las observaciones pendientes de un neumático que pueden ser solucionadas por un tipo de movimiento específico
      operationId: getObservacionesSolucionables
      parameters:
        - name: neumaticoId
          in: path
          required: true
          description: ID del neumático
          schema:
            type: integer
            format: int32
            minimum: 1
          example: 1
        - name: tipoMovimientoId
          in: query
          required: true
          description: ID del tipo de movimiento
          schema:
            type: integer
            format: int32
            minimum: 1
          example: 2
      responses:
        '200':
          description: Lista de observaciones solucionables obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ObservacionNeumaticoResponse'
              examples:
                success:
                  summary: Respuesta exitosa con observaciones
                  value:
                    - id: 1
                      neumaticoSummary:
                        id: 1
                        serieCodigo: "NEU-001-2023"
                      equipoId: 123
                      posicion: 1
                      tipoObservacionResponse:
                        id: 1
                        nombre: "Desgaste irregular"
                        ambito: "Rodadura"
                        descripcion: "Desgaste no uniforme en la banda de rodadura"
                        activo: true
                      descripcion: "Se observa desgaste irregular en el hombro derecho"
                      estadoObservacionResponse:
                        id: 1
                        nombre: "Pendiente"
                        descripcion: "Observación pendiente de resolución"
                      fechaCreacion: "2023-08-01 10:30:00"
                      userInfoResponseCreacion:
                        id: 5
                        name: "Carlos"
                        lastName: "González"
                        fullName: "Carlos González"
                      fechaResolucion: null
                      usuarioResolucionId: null
                      comentarioResolucion: null
                empty:
                  summary: No hay observaciones solucionables
                  value: []
        '400':
          description: Parámetros de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "TYR-VAL-001"
                message: "neumaticoId debe ser un número positivo válido"
                path: "/api/v1/observaciones-neumaticos/neumatico/0/solucionables"
                timestamp: "2023-08-06 10:30:00"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "TYR-OBS-OPE-001"
                message: "Error al obtener observaciones solucionables"
                path: "/api/v1/observaciones-neumaticos/neumatico/1/solucionables"
                timestamp: "2023-08-06 10:30:00"
  /api/v1/observaciones-neumaticos/neumatico/{neumaticoId}:
    get:
      tags:
        - observaciones
      summary: Obtener observaciones de un neumático
      description: Retorna todas las observaciones o solo las pendientes de un neumático específico
      operationId: getObservacionesByNeumatico
      parameters:
        - name: neumaticoId
          in: path
          required: true
          description: ID del neumático
          schema:
            type: integer
            format: int32
            minimum: 1
          example: 1
        - name: estado
          in: query
          required: false
          description: Filtro por estado de observación. Si se especifica "pendiente", solo retorna observaciones pendientes
          schema:
            type: string
            enum: [pendiente]
          example: pendiente
      responses:
        '200':
          description: Lista de observaciones obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ObservacionNeumaticoResponse'
              examples:
                all_observations:
                  summary: Todas las observaciones
                  value:
                    - id: 1
                      neumaticoSummary:
                        id: 1
                        serieCodigo: "NEU-001-2023"
                      equipoId: 123
                      posicion: 1
                      tipoObservacionResponse:
                        id: 1
                        nombre: "Desgaste irregular"
                        ambito: "Rodadura"
                        descripcion: "Desgaste no uniforme en la banda de rodadura"
                        activo: true
                      descripcion: "Se observa desgaste irregular en el hombro derecho"
                      estadoObservacionResponse:
                        id: 2
                        nombre: "Resuelto"
                        descripcion: "Observación resuelta satisfactoriamente"
                      fechaCreacion: "2023-08-01 10:30:00"
                      userInfoResponseCreacion:
                        id: 5
                        name: "Carlos"
                        lastName: "González"
                        fullName: "Carlos González"
                      fechaResolucion: "2023-08-02 14:15:00"
                      usuarioResolucionId: 3
                      comentarioResolucion: "Se realizó rotación del neumático"
                    - id: 2
                      neumaticoSummary:
                        id: 1
                        serieCodigo: "NEU-001-2023"
                      equipoId: 123
                      posicion: 1
                      tipoObservacionResponse:
                        id: 2
                        nombre: "Presión baja"
                        ambito: "Inflado"
                        descripcion: "Presión por debajo del rango recomendado"
                        activo: true
                      descripcion: "Presión detectada en 80 PSI"
                      estadoObservacionResponse:
                        id: 1
                        nombre: "Pendiente"
                        descripcion: "Observación pendiente de resolución"
                      fechaCreacion: "2023-08-05 09:20:00"
                      userInfoResponseCreacion:
                        id: 4
                        name: "Ana"
                        lastName: "Martínez"
                        fullName: "Ana Martínez"
                      fechaResolucion: null
                      usuarioResolucionId: null
                      comentarioResolucion: null
                pending_only:
                  summary: Solo observaciones pendientes
                  value:
                    - id: 2
                      neumaticoSummary:
                        id: 1
                        serieCodigo: "NEU-001-2023"
                      equipoId: 123
                      posicion: 1
                      tipoObservacionResponse:
                        id: 2
                        nombre: "Presión baja"
                        ambito: "Inflado"
                        descripcion: "Presión por debajo del rango recomendado"
                        activo: true
                      descripcion: "Presión detectada en 80 PSI"
                      estadoObservacionResponse:
                        id: 1
                        nombre: "Pendiente"
                        descripcion: "Observación pendiente de resolución"
                      fechaCreacion: "2023-08-05 09:20:00"
                      userInfoResponseCreacion:
                        id: 4
                        name: "Ana"
                        lastName: "Martínez"
                        fullName: "Ana Martínez"
                      fechaResolucion: null
                      usuarioResolucionId: null
                      comentarioResolucion: null
                empty:
                  summary: No hay observaciones
                  value: []
        '400':
          description: Parámetros de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "TYR-VAL-001"
                message: "neumaticoId debe ser un número positivo válido"
                path: "/api/v1/observaciones-neumaticos/neumatico/0"
                timestamp: "2023-08-06 10:30:00"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "TYR-OBS-OPE-002"
                message: "Error al obtener observaciones del neumático"
                path: "/api/v1/observaciones-neumaticos/neumatico/1"
                timestamp: "2023-08-06 10:30:00"
  /api/v1/observaciones-neumaticos/equipo/{equipoId}/pendientes:
    get:
      tags:
        - observaciones
      summary: Obtener observaciones pendientes por equipo
      description: Retorna todas las observaciones pendientes asociadas a un equipo específico
      operationId: getObservacionesPendientesByEquipo
      parameters:
        - name: equipoId
          in: path
          required: true
          description: ID del equipo
          schema:
            type: integer
            format: int32
            minimum: 1
          example: 123
      responses:
        '200':
          description: Lista de observaciones pendientes obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ObservacionNeumaticoResponse'
              examples:
                success:
                  summary: Respuesta exitosa con observaciones pendientes
                  value:
                    - id: 1
                      neumaticoSummary:
                        id: 1
                        serieCodigo: "NEU-001-2023"
                      equipoId: 123
                      posicion: 1
                      tipoObservacionResponse:
                        id: 1
                        nombre: "Desgaste irregular"
                        ambito: "Rodadura"
                        descripcion: "Desgaste no uniforme en la banda de rodadura"
                        activo: true
                      descripcion: "Se observa desgaste irregular en el hombro derecho"
                      estadoObservacionResponse:
                        id: 1
                        nombre: "Pendiente"
                        descripcion: "Observación pendiente de resolución"
                      fechaCreacion: "2023-08-01 10:30:00"
                      userInfoResponseCreacion:
                        id: 5
                        name: "Carlos"
                        lastName: "González"
                        fullName: "Carlos González"
                      fechaResolucion: null
                      usuarioResolucionId: null
                      comentarioResolucion: null
                    - id: 3
                      neumaticoSummary:
                        id: 2
                        serieCodigo: "NEU-002-2023"
                      equipoId: 123
                      posicion: 2
                      tipoObservacionResponse:
                        id: 2
                        nombre: "Presión baja"
                        ambito: "Inflado"
                        descripcion: "Presión por debajo del rango recomendado"
                        activo: true
                      descripcion: "Presión detectada en 75 PSI"
                      estadoObservacionResponse:
                        id: 1
                        nombre: "Pendiente"
                        descripcion: "Observación pendiente de resolución"
                      fechaCreacion: "2023-08-05 14:30:00"
                      userInfoResponseCreacion:
                        id: 6
                        name: "Luis"
                        lastName: "Rodríguez"
                        fullName: "Luis Rodríguez"
                      fechaResolucion: null
                      usuarioResolucionId: null
                      comentarioResolucion: null
                empty:
                  summary: No hay observaciones pendientes para el equipo
                  value: []
        '400':
          description: Parámetros de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "TYR-VAL-001"
                message: "equipoId debe ser un número positivo válido"
                path: "/api/v1/observaciones-neumaticos/equipo/0/pendientes"
                timestamp: "2023-08-06 10:30:00"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "TYR-OBS-OPE-003"
                message: "Error al obtener observaciones pendientes del equipo"
                path: "/api/v1/observaciones-neumaticos/equipo/123/pendientes"
                timestamp: "2023-08-06 10:30:00"
  /api/v1/observaciones-neumaticos:
    post:
      tags:
        - observaciones
      summary: Crear nueva observación
      description: Crea una nueva observación para un neumático
      operationId: createObservacionNeumatico
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservacionNeumaticoNuevoRequest'
            example:
              neumaticoId: 1
              equipoId: 123
              posicion: 1
              tipoObservacionId: 1
              descripcion: "Se observa desgaste irregular en el hombro derecho del neumático"
              usuarioCreacionId: 5
      responses:
        '201':
          description: Observación creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservacionNeumaticoResponse'
              example:
                id: 3
                neumaticoSummary:
                  id: 1
                  serieCodigo: "NEU-001-2023"
                equipoId: 123
                posicion: 1
                tipoObservacionResponse:
                  id: 1
                  nombre: "Desgaste irregular"
                  ambito: "Rodadura"
                  descripcion: "Desgaste no uniforme en la banda de rodadura"
                  activo: true
                descripcion: "Se observa desgaste irregular en el hombro derecho del neumático"
                estadoObservacionResponse:
                  id: 1
                  nombre: "Pendiente"
                  descripcion: "Observación pendiente de resolución"
                fechaCreacion: "2023-08-06 10:30:00"
                userInfoResponseCreacion:
                  id: 5
                  name: "Carlos"
                  lastName: "González"
                  fullName: "Carlos González"
                fechaResolucion: null
                usuarioResolucionId: null
                comentarioResolucion: null
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "TYR-VAL-001"
                message: "descripcion es requerida y no puede estar vacía"
                path: "/api/v1/observaciones-neumaticos"
                timestamp: "2023-08-06 10:30:00"
        '404':
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "TYR-OBS-NFE-001"
                message: "Tipo de observación con ID 999 no encontrado"
                path: "/api/v1/observaciones-neumaticos"
                timestamp: "2023-08-06 10:30:00"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "TYR-OBS-CRE-001"
                message: "Error al crear observación"
                path: "/api/v1/observaciones-neumaticos"
                timestamp: "2023-08-06 10:30:00"
  /api/v1/observaciones-neumaticos/{observacionId}:
    patch:
      tags:
        - observaciones
      summary: Actualizar observación
      description: Actualiza el estado, usuario de resolución y/o comentarios de una observación existente
      operationId: updateObservacionNeumatico
      parameters:
        - name: observacionId
          in: path
          required: true
          description: ID de la observación a actualizar
          schema:
            type: integer
            format: int32
            minimum: 1
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservacionNeumaticoUpdateRequest'
            examples:
              resolve_observation:
                summary: Resolver observación
                value:
                  estadoObservacionId: 2
                  usuarioResolucionId: 3
                  comentarioResolucion: "Se realizó rotación del neumático para corregir el desgaste irregular"
              add_comment_only:
                summary: Agregar solo comentario
                value:
                  comentarioResolucion: "En seguimiento - revisar en próxima inspección"
      responses:
        '200':
          description: Observación actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservacionNeumaticoResponse'
              example:
                id: 1
                neumaticoSummary:
                  id: 1
                  serieCodigo: "NEU-001-2023"
                equipoId: 123
                posicion: 1
                tipoObservacionResponse:
                  id: 1
                  nombre: "Desgaste irregular"
                  ambito: "Rodadura"
                  descripcion: "Desgaste no uniforme en la banda de rodadura"
                  activo: true
                descripcion: "Se observa desgaste irregular en el hombro derecho"
                estadoObservacionResponse:
                  id: 2
                  nombre: "Resuelto"
                  descripcion: "Observación resuelta satisfactoriamente"
                fechaCreacion: "2023-08-01 10:30:00"
                userInfoResponseCreacion:
                  id: 5
                  name: "Carlos"
                  lastName: "González"
                  fullName: "Carlos González"
                fechaResolucion: "2023-08-06 10:30:00"
                usuarioResolucionId: 3
                comentarioResolucion: "Se realizó rotación del neumático para corregir el desgaste irregular"
        '400':
          description: Datos de entrada inválidos o reglas de negocio violadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_error:
                  summary: Error de validación
                  value:
                    code: "TYR-VAL-001"
                    message: "Debe proporcionar al menos un campo para actualizar"
                    path: "/api/v1/observaciones-neumaticos/1"
                    timestamp: "2023-08-06 10:30:00"
                business_rule_error:
                  summary: Error de regla de negocio
                  value:
                    code: "TYR-OBS-UPD-002"
                    message: "No se puede cambiar el estado de una observación ya resuelta"
                    path: "/api/v1/observaciones-neumaticos/1"
                    timestamp: "2023-08-06 10:30:00"
        '404':
          description: Observación no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "TYR-OBS-NFE-001"
                message: "Observación con ID 999 no encontrada"
                path: "/api/v1/observaciones-neumaticos/999"
                timestamp: "2023-08-06 10:30:00"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "TYR-OBS-UPD-001"
                message: "Error al actualizar observación"
                path: "/api/v1/observaciones-neumaticos/1"
                timestamp: "2023-08-06 10:30:00"
...
components:
  schemas:
...
    ObservacionNeumaticoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: ID único de la observación
          example: 1
        neumaticoSummary:
          $ref: '#/components/schemas/NeumaticoSummaryResponse'
        equipoId:
          type: integer
          format: int32
          description: ID del equipo
          example: 123
        posicion:
          type: integer
          format: int32
          description: Posición del neumático en el equipo
          example: 1
        tipoObservacionResponse:
          $ref: '#/components/schemas/TipoObservacionResponse'
        descripcion:
          type: string
          description: Descripción detallada de la observación
          example: "Se observa desgaste irregular en el hombro derecho"
        estadoObservacionResponse:
          $ref: '#/components/schemas/EstadoObservacionResponse'
        fechaCreacion:
          type: string
          format: date-time
          description: Fecha y hora de creación
          example: "2023-08-01 10:30:00"
        userInfoResponseCreacion:
          $ref: '#/components/schemas/UserInfoResponse'
        fechaResolucion:
          type: string
          format: date-time
          nullable: true
          description: Fecha y hora de resolución
          example: "2023-08-02 14:15:00"
        usuarioResolucionId:
          type: integer
          format: int32
          nullable: true
          description: ID del usuario que resolvió la observación
          example: 3
        comentarioResolucion:
          type: string
          nullable: true
          description: Comentarios de resolución
          example: "Se realizó rotación del neumático"
    TipoObservacionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 1
        nombre:
          type: string
          example: "Desgaste irregular"
        ambito:
          type: string
          example: "Rodadura"
        descripcion:
          type: string
          example: "Desgaste no uniforme en la banda de rodadura"
        activo:
          type: boolean
          example: true
    EstadoObservacionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 1
        nombre:
          type: string
          example: "Pendiente"
        descripcion:
          type: string
          example: "Observación pendiente de resolución"
    NeumaticoSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: Identificador único del neumático
          example: 1
        serieCodigo:
          type: string
          description: Código de serie del neumático
          example: "NEU-001-2023"
    UserInfoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 1
        name:
          type: string
          example: "Juan"
        lastName:
          type: string
          example: "Pérez"
        fullName:
          type: string
          example: "Juan Pérez"
    ObservacionNeumaticoNuevoRequest:
      type: object
      required:
        - neumaticoId
        - equipoId
        - posicion
        - tipoObservacionId
        - descripcion
        - usuarioCreacionId
      properties:
        neumaticoId:
          type: integer
          format: int32
          minimum: 1
          description: ID del neumático
          example: 1
        equipoId:
          type: integer
          format: int32
          minimum: 1
          description: ID del equipo
          example: 123
        posicion:
          type: integer
          format: int32
          minimum: 1
          description: Posición del neumático en el equipo
          example: 1
        tipoObservacionId:
          type: integer
          format: int32
          minimum: 1
          description: ID del tipo de observación
          example: 1
        descripcion:
          type: string
          minLength: 1
          maxLength: 1000
          description: Descripción detallada de la observación
          example: "Se observa desgaste irregular en el hombro derecho del neumático"
        usuarioCreacionId:
          type: integer
          format: int32
          minimum: 1
          description: ID del usuario que crea la observación
          example: 5
    ObservacionNeumaticoUpdateRequest:
      type: object
      properties:
        estadoObservacionId:
          type: integer
          format: int32
          minimum: 1
          description: ID del nuevo estado de la observación
          example: 2
        usuarioResolucionId:
          type: integer
          format: int32
          minimum: 1
          description: ID del usuario que resuelve la observación
          example: 3
        comentarioResolucion:
          type: string
          maxLength: 1000
          description: Comentarios sobre la resolución
          example: "Se realizó rotación del neumático para corregir el desgaste irregular"
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Código de error único
          example: "TYR-VAL-001"
        message:
          type: string
          description: Mensaje descriptivo del error
          example: "Parámetro inválido"
        path:
          type: string
          description: Ruta del endpoint donde ocurrió el error
          example: "/api/v1/neumaticos/equipo/123"
        timestamp:
          type: string
          format: date-time
          description: Fecha y hora del error
          example: "2023-08-06 10:30:00"
...